move qt tests to qt/vispy module
