+++ src/napari/utils/_tests/test_camera_orientations.py
@@ -30,54 +30,6 @@ ANGLES = st.tuples(
 )
 
 
-@pytest.fixture(scope='module')
-def arcball_camera():
-    from vispy import scene
-
-    canvas = scene.SceneCanvas(size=(100, 100), show=False)
-    try:
-        view = canvas.central_widget.add_view()
-        camera = scene.ArcballCamera(fov=0)
-        view.camera = camera
-        camera.set_range(x=(0, 10), y=(0, 10), z=(0, 10))
-        yield camera
-    finally:
-        canvas.close()
-
-
-@pytest.mark.filterwarnings('ignore:gimbal lock')
-@pytest.mark.parametrize('orientation', ORIENTATIONS)
-@given(angles=ANGLES)
-@settings(max_examples=20, deadline=None)
-def test_view_direction_correct_under_rotation(
-    orientation, angles, arcball_camera
-):
-    """Check that the napari direction math matches a real VisPy 3D camera."""
-    from napari._vispy.camera import (
-        _get_vispy_flipped_axes,
-        napari_angles_to_vispy_quat,
-    )
-
-    camera = arcball_camera
-    camera.flip = _get_vispy_flipped_axes(orientation, ndisplay=3)
-    camera.set_state(
-        _quaternion=napari_angles_to_vispy_quat(angles, orientation)
-    )
-    camera.view_changed()
-    matrix_inv = np.linalg.inv(camera.transform.matrix[:3, :3])
-    # VisPy uses xyz coordinates; napari uses zyx, hence the reversal.
-    view_direction = (-matrix_inv[:, 2])[::-1]
-    up_direction = (matrix_inv[:, 1])[::-1]
-    assert np.allclose(
-        view_direction_from_angles(angles, orientation),
-        view_direction,
-    )
-    assert np.allclose(
-        up_direction_from_angles(angles, orientation),
-        up_direction,
-    )
-
-
 @pytest.mark.parametrize('orientation', ORIENTATIONS)
 def test_home_view_matches_base_directions(orientation):
     """With zero angles the camera must show the home view."""
@@ -121,28 +73,6 @@ def test_single_angle_rotation_axis(angle_index, axis_component):
     axis = Rotation.from_matrix(rotation).as_rotvec()
     axis /= np.linalg.norm(axis)
     assert np.allclose(np.abs(axis), np.eye(3)[axis_component])
-
-
-@pytest.mark.parametrize('orientation', ORIENTATIONS)
-@given(angles=ANGLES)
-@settings(max_examples=50, deadline=None)
-def test_vispy_quat_roundtrip(orientation, angles):
-    """Check the vispy quaternion conversion inverts exactly."""
-    from napari._vispy.camera import (
-        napari_angles_to_vispy_quat,
-        vispy_quat_to_napari_angles,
-    )
-
-    quat = napari_angles_to_vispy_quat(angles, orientation)
-    recovered = vispy_quat_to_napari_angles(quat, orientation)
-    assert np.allclose(
-        view_direction_from_angles(recovered, orientation),
-        view_direction_from_angles(angles, orientation),
-    )
-    assert np.allclose(
-        up_direction_from_angles(recovered, orientation),
-        up_direction_from_angles(angles, orientation),
-    )
 
 
 @pytest.mark.parametrize('orientation', ORIENTATIONS)
